src/requirements.txt
pandas
numpy
openpyxl
dask
dask[dataframe]
sqlalchemy
memory_profiler
psycopg2
python-dotenv
